requirements.txt
albumentations==2.0.8
kaggle==1.6.17
numpy==2.3.0
opencv_python_headless==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
scikit_learn==1.7.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
